+++ projects/hipcub/benchmark/benchmark_block_exchange.cpp
@@ -288,7 +288,7 @@ void add_benchmarks(primbench::executor& executor)
 
     BENCHMARK_TYPE(int, 256);
     BENCHMARK_TYPE(int8_t, 256);
-    BENCHMARK_TYPE(long long, 256);
+    BENCHMARK_TYPE(int64_t, 256);
     BENCHMARK_TYPE(custom_float2, 256);
     BENCHMARK_TYPE(custom_double2, 256);
 }
