+++ projects/hipcub/benchmark/benchmark_block_merge_sort.cpp
@@ -198,10 +198,10 @@ void add_benchmarks(primbench::executor& executor)
     BENCHMARK_TYPE(uint8_t, 256);
     BENCHMARK_TYPE(uint8_t, 512);
 
-    BENCHMARK_TYPE(long long, 64);
-    BENCHMARK_TYPE(long long, 128);
-    BENCHMARK_TYPE(long long, 256);
-    BENCHMARK_TYPE(long long, 512);
+    BENCHMARK_TYPE(int64_t, 64);
+    BENCHMARK_TYPE(int64_t, 128);
+    BENCHMARK_TYPE(int64_t, 256);
+    BENCHMARK_TYPE(int64_t, 512);
 }
 
 int main(int argc, char* argv[])
